refactor: better structure spring commons
