+++ .config/detekt/detekt.yaml
@@ -38,9 +38,11 @@ ktlint:
   code_style: ktlint_official
   Indentation:
     indentStyle: tab
-  FunctionSignature:  # Detekt overrides this with `default` instead of the (better) default of Ktlint
-    active: false  # https://github.com/detekt/detekt/issues/9295
+  FunctionSignature:  # Detekt overrides this with `default` instead of the (better) default of Ktlint.
+    active: false  # https://github.com/detekt/detekt/issues/9295 -- remove on 2.0.0-alpha.4+
     functionBodyExpressionWrapping: multiline
+  ValueParameterComment:
+    active: false  # Annoying when adding comments after annotations. Wouldn't pass code review anyway.
 
 performance:
   CouldBeSequence:
