+++ commons/spring/src/main/kotlin/org/elaastix/commons/jpa/entity/EntityListener.kt
@@ -17,22 +17,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.elaastix.commons.jpa.hibernate
+package org.elaastix.commons.jpa.entity
 
-import kotlinx.serialization.json.Json
-import org.hibernate.cfg.AvailableSettings
-import org.springframework.boot.autoconfigure.AutoConfiguration
-import org.springframework.boot.hibernate.autoconfigure.HibernatePropertiesCustomizer
-import org.springframework.context.annotation.Bean
+import jakarta.persistence.PrePersist
+import org.elaastix.commons.platform.JpaImmutable
+import kotlin.time.Clock
 
-/**
- * Autoconfiguration class loading the library's Hibernate default settings.
- */
-@AutoConfiguration
-class ElaastixHibernateAutoConfiguration {
-	@Bean
-	@Suppress("UndocumentedPublicFunction")
-	fun jsonFormatMapperCustomizer(json: Json): HibernatePropertiesCustomizer = {
-		it[AvailableSettings.JSON_FORMAT_MAPPER] = HibernateKotlinxJsonMapper(json)
+internal class EntityListener {
+	@PrePersist
+	fun prePersist(entity: AbstractEntity) {
+		@OptIn(JpaImmutable::class)
+		entity.updatedAt = Clock.System.now()
 	}
 }
