+++ commons/spring/src/main/kotlin/org/elaastix/commons/jpa/repository/ElaastixRepositoryImpl.kt
@@ -17,11 +17,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.elaastix.commons.jpa
+package org.elaastix.commons.jpa.repository
 
 import io.hypersistence.utils.spring.repository.BaseJpaRepositoryImpl
 import jakarta.persistence.EntityManager
 import org.elaastix.commons.data.Uuid
+import org.elaastix.commons.jpa.entity.AbstractEntity
 import org.springframework.data.jpa.repository.support.JpaEntityInformation
 
 /**
