+++ commons/spring/src/main/resources/META-INF/services/org.hibernate.boot.model.TypeContributor
@@ -1,1 +1,1 @@
-org.elaastix.commons.jpa.HibernateTypeContributor
+org.elaastix.commons.jpa.hibernate.ElaastixTypeContributor
