+++ commons/spring/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,3 @@
-org.elaastix.commons.jpa.hibernate.ElaastixHibernateAutoConfiguration
-org.elaastix.commons.conversion.SpringConvertersAutoConfiguration
-org.elaastix.commons.openapi.SpringdocAutoConfiguration
+org.elaastix.commons.boot.autoconfigure.ElaastixHibernateAutoConfiguration
+org.elaastix.commons.boot.autoconfigure.SpringConvertersAutoConfiguration
+org.elaastix.commons.boot.autoconfigure.SpringdocAutoConfiguration
