+++ server/src/main/kotlin/org/elaastix/server/ElaastixServerApplication.kt
@@ -23,7 +23,7 @@ import io.swagger.v3.oas.annotations.OpenAPIDefinition
 import io.swagger.v3.oas.annotations.info.Contact
 import io.swagger.v3.oas.annotations.info.Info
 import io.swagger.v3.oas.annotations.info.License
-import org.elaastix.commons.jpa.ElaastixRepositoryImpl
+import org.elaastix.commons.jpa.repository.ElaastixRepositoryImpl
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
