+++ server/src/main/kotlin/org/elaastix/server/activities/response/entities/ClosedQuestionEntity.kt
@@ -47,6 +47,7 @@ class ClosedQuestionEntity(
 
 	/** The expected answer. */
 	@JdbcTypeCode(SqlTypes.JSON)
+	@Suppress("JpaAttributeTypeInspection") // https://youtrack.jetbrains.com/issue/IDEA-191568
 	var expectedAnswer: ClosedAnswer,
 
 	answerExplanation: RichContent?,
