+++ server/src/main/kotlin/org/elaastix/server/activities/response/entities/ClosedResponseEntity.kt
@@ -36,6 +36,7 @@ class ClosedResponseEntity(
 
 	/** The answer given by the learner. May be empty, but is required. */
 	@JdbcTypeCode(SqlTypes.JSON)
+	@Suppress("JpaAttributeTypeInspection") // https://youtrack.jetbrains.com/issue/IDEA-191568
 	var answer: ClosedAnswer,
 
 	selfExplanation: FormattedContent?,
