+++ server/src/main/kotlin/org/elaastix/server/activities/response/entities/OpenQuestionEntity.kt
@@ -43,6 +43,7 @@ class OpenQuestionEntity(
 	 * question author to put themselves in the student's shoes while preparing the questions.
 	 */
 	@JdbcTypeCode(SqlTypes.JSON)
+	@Suppress("JpaAttributeTypeInspection") // https://youtrack.jetbrains.com/issue/IDEA-191568
 	var expectedAnswer: FormattedContent,
 
 	answerExplanation: RichContent?,
