+++ server/src/main/kotlin/org/elaastix/server/activities/response/entities/OpenResponseEntity.kt
@@ -35,6 +35,7 @@ class OpenResponseEntity(
 
 	/** The answer given by the learner. May be empty, but is required. */
 	@JdbcTypeCode(SqlTypes.JSON)
+	@Suppress("JpaAttributeTypeInspection") // https://youtrack.jetbrains.com/issue/IDEA-191568
 	var answer: FormattedContent,
 
 	selfExplanation: FormattedContent?,
