+++ server/src/main/kotlin/org/elaastix/server/activities/response/entities/QuestionEntity.kt
@@ -23,7 +23,7 @@ import jakarta.persistence.Entity
 import jakarta.persistence.Inheritance
 import jakarta.persistence.InheritanceType
 import jakarta.persistence.ManyToOne
-import org.elaastix.commons.jpa.AbstractEntity
+import org.elaastix.commons.jpa.entity.AbstractEntity
 import org.elaastix.mm.activity.Material
 import org.elaastix.mm.content.RichContent
 import org.elaastix.server.users.entities.UserEntity
@@ -38,10 +38,12 @@ import org.hibernate.type.SqlTypes
 abstract class QuestionEntity(
 	/** The question's statement. */
 	@JdbcTypeCode(SqlTypes.JSON)
+	@Suppress("JpaAttributeTypeInspection") // https://youtrack.jetbrains.com/issue/IDEA-191568
 	var statement: RichContent,
 
 	/** The explanation of the expected answer. Optional. */
 	@JdbcTypeCode(SqlTypes.JSON)
+	@Suppress("JpaAttributeTypeInspection") // https://youtrack.jetbrains.com/issue/IDEA-191568
 	var answerExplanation: RichContent?,
 
 	@ManyToOne
